requirements.txt
anyio==3.7.0
astroid==2.15.5
autopep8==2.0.2
colorama==0.4.6
dill==0.3.6
dnspython==2.3.0
fastapi==0.98.0
flake8==6.0.0
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy==1.3.0
mypy-extensions==1.0.0
platformdirs==3.6.0
pycodestyle==2.10.0
pydantic==1.10.9
pyflakes==3.0.1
pylint==2.17.4
pymongo==4.4.0
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
tomlkit==0.11.8
typing_extensions==4.6.3
validate-email==1.3
wrapt==1.15.0

h11~=0.14.0
pip~=23.1.2
PyYAML~=6.0
setuptools~=65.5.0
click~=8.1.3
websockets~=11.0.3
uvicorn~=0.22.0
httptools~=0.5.0
watchfiles~=0.19.0
python-multipart~=0.0.6
passlib~=1.7.4
PyJWT~=2.7.0